requirements.txt
# python 3.11
torch
torch-geometric
wandb
pyyaml
